Rename package to bmiheat
