+++ src/main/java/edu/colorado/csdms/bmiheat/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Model the diffusion of heat over a 2D plate.
  */
-package edu.colorado.csdms.heat;
+package edu.colorado.csdms.bmiheat;
